+++ src/test/kotlin/com/zepben/auth/server/JWTAuthenticatorTest.kt
@@ -20,6 +20,7 @@ import com.auth0.jwk.Jwk
 import com.auth0.jwk.JwkException
 import com.auth0.jwk.UrlJwkProvider
 import com.auth0.jwt.exceptions.*
+import com.zepben.auth.client.ProviderDetails
 import com.zepben.auth.common.StatusCode
 import com.zepben.auth.server.JWTAuthoriser.authorise
 import com.zepben.testutils.auth.*
@@ -35,7 +36,12 @@ class JWTAuthenticatorTest {
 
     @Test
     fun testAuth() {
-        var ta = JWTAuthenticator("https://fake-aud/", "https://issuer/", MockJwksUrlProvider())
+        var ta = JWTAuthenticator("https://fake-aud/", listOf(TrustedIssuer("https://issuer/", ProviderDetails("dunno", "https://whereileftmy/keys/"))),
+            verifierBuilder = JWTMultiIssuerVerifierBuilder(
+                "https://fake-aud/",
+                listOf(TrustedIssuer("https://issuer/", ProviderDetails("dunno", "https://whereileft.my/keys/"))),
+                JWKHolder { _ -> MockJwksUrlProvider().all.associateBy { it.id } } ))
+
         var authResp = ta.authenticate(TOKEN)
         assertThat(authResp.statusCode, equalTo(StatusCode.OK))
         val successfulToken = authResp.token!!
@@ -44,7 +50,7 @@ class JWTAuthenticatorTest {
 
         authResp = authorise(successfulToken, "bacon")
         assertThat(authResp.statusCode, equalTo(StatusCode.UNAUTHENTICATED))
-        assertThat(authResp.message, equalTo("Token was missing required claim bacon"))
+        assertThat(authResp.message, equalTo("Token was missing a required claim. Had [read:network, read:ewb, write:metrics, write:network] but needed [bacon]"))
 
         authResp = ta.authenticate("broken")
         assertThat(authResp.statusCode, equalTo(StatusCode.UNAUTHENTICATED))
@@ -62,17 +68,27 @@ class JWTAuthenticatorTest {
         assertThat(authResp.statusCode, equalTo(StatusCode.UNAUTHENTICATED))
         assertThat(authResp.cause, instanceOf(TokenExpiredException::class.java))
 
-        ta = JWTAuthenticator("https://wrong-aud/", "https://issuer/", MockJwksUrlProvider())
+        ta = JWTAuthenticator("https://wrong-aud/", listOf(TrustedIssuer("https://issuer/", ProviderDetails("dunno", "https://whereileftmy/keys/"))),
+            verifierBuilder = JWTMultiIssuerVerifierBuilder(
+                "https://wrong-aud/",
+                listOf(TrustedIssuer("https://issuer/", ProviderDetails("dunno", "https://whereileft.my/keys/"))),
+                JWKHolder { _ -> MockJwksUrlProvider().all.associateBy { it.id } } ))
+
         authResp = ta.authenticate(TOKEN)
         assertThat(authResp.statusCode, equalTo(StatusCode.PERMISSION_DENIED))
         assertThat(authResp.cause, instanceOf(InvalidClaimException::class.java))
         assertThat(authResp.message, equalTo("The Claim 'aud' value doesn't contain the required audience."))
 
-        ta = JWTAuthenticator("https://fake-aud/", "wrong-issuer", MockJwksUrlProvider())
+        ta = JWTAuthenticator("https://fake-aud/", listOf(TrustedIssuer("wrong-issuer", ProviderDetails("dunno", "https://whereileftmy/keys/"))),
+            verifierBuilder = JWTMultiIssuerVerifierBuilder(
+                "https://fake-aud/",
+                listOf(TrustedIssuer("wrong-issuer", ProviderDetails("dunno", "https://whereileft.my/keys/"))),
+                JWKHolder { _ -> MockJwksUrlProvider().all.associateBy { it.id } } ))
+
         authResp = ta.authenticate(TOKEN)
         assertThat(authResp.statusCode, equalTo(StatusCode.PERMISSION_DENIED))
         assertThat(authResp.cause, instanceOf(InvalidClaimException::class.java))
-        assertThat(authResp.message, equalTo("The Claim 'iss' value doesn't match the required issuer."))
+        assertThat(authResp.message, equalTo("Unknown or untrusted issuer: https://issuer/"))
     }
 
     @Test
@@ -81,12 +97,13 @@ class JWTAuthenticatorTest {
         val mockJWK = mockk<UrlJwkProvider> {
             every { all } returns listOf(jwk)
         }
-        val ta = JWTAuthenticator("https://fake-aud/", "https://issuer/", mockJWK)
 
-        assertThat(ta.getKeyFromJwk("fakekid"), equalTo(jwk))
+        val ta = JWTMultiIssuerVerifierBuilder("https://fake-aud/", listOf(TrustedIssuer("https://issuer/", ProviderDetails("dunno", "https://whereileftmy/keys/")))
+        , JWKHolder { _ -> mockJWK.all.associateBy { it.id } } )
+        assertThat(ta.jwkHolder.getKeyFromJwk("fakekid", TrustedIssuer("ignored_for_now", ProviderDetails("",""))), equalTo(jwk))
 
         expect {
-            ta.getKeyFromJwk("fakekey")
+            ta.jwkHolder.getKeyFromJwk("fakekey", TrustedIssuer("ignored_for_now", ProviderDetails("","")))
         }.toThrow(JwkException::class.java)
             .withMessage("Unable to find key fakekey in jwk endpoint. Check your JWK URL.")
     }